Changed a way to query codec element.
The query routine that copies some data from device is called when each element is used.
However, the codec information is static data and it does not
need to copy data from device everytime. So, the data is cached in the driver and
this module gets them from driver, not device.

Change-Id: I31ee17b508ed859297bfa25b09171858fdc37329
Signed-off-by: Kitae Kim <kt920.kim@samsung.com>