fix a sync problem.
There is a variable to manage how many codecs are used in a process.
As depending on the count variable, this plugin determine whether release resources or not.
However, there is an exception case that the variable can be invalid,
and it causes SEGV problem when opening and closing CODECs concurrently.

Change-Id: If5bab4c679f986d0cb7108fd36c23ce57c741ab2
Signed-off-by: Kitae Kim <kt920.kim@samsung.com>